--- a/Content.Server/Atmos/EntitySystems/AtmosphereSystem.API.cs
+++ b/Content.Server/Atmos/EntitySystems/AtmosphereSystem.API.cs
@@ -355,9 +355,7 @@ public partial class AtmosphereSystem
         grid.Comp.DeltaPressureEntityLookup[ent.Owner] = grid.Comp.DeltaPressureEntities.Count;
         grid.Comp.DeltaPressureEntities.Add(ent);
 
-        ent.Comp.CurrentPosition = _map.CoordinatesToTile(grid,
-            Comp<MapGridComponent>(grid),
-            xform.Coordinates);
+        CacheAirtightStructureOffsets(ent, Comp<MapGridComponent>(grid), xform);
 
         ent.Comp.GridUid = grid.Owner;
         ent.Comp.InProcessingList = true;